Rewrote config for i3

+++ install.sh
@@ -9,14 +9,19 @@ git submodule update
 echo "Downloading and installing oh-my-zsh"
 wget --quiet --no-check-certificate https://github.com/robbyrussell/oh-my-zsh/raw/master/tools/install.sh -O - | sh
 
-echo "Downloading fasd"
-git clone --quiet https://github.com/clvv/fasd.git
 
-echo "Installing fasd"
-cd fasd
-make install --quiet
-cd ..
-rm -r fasd/
+if ! [ "$(exists fasd)" -eq 1 ]
+then
+	echo "Downloading fasd"
+	git clone --quiet https://github.com/clvv/fasd.git
+
+	echo "Setting up FASD - a command-line productivity booster"
+	cd fasd
+	make install --quiet
+	cd ..
+	rm -r fasd/
+fi
+
 
 echo "Deleting the old files"
 mv ~/.zshrc ~/.zshrc_old
@@ -26,8 +31,8 @@ mv ~/.gitignore ~/.gitignore_old
 mv ~/.tmux.conf ~/.tmux.conf_old
 mv ~/.gemrc ~/.gemrc_old
 mv ~/.bash_profile ~/.bash_profile_old
-mv ~/.config/i3/config ~/.config/i3/config_old
-mv ~/.config/i3/status ~/.config/i3/status_old
+mv ~/.i3/config ~/.i3/config_old
+mv ~/.config/i3status/config ~/.config/i3status/config_old
 
 echo "Symlinking files"
 ln -s $scriptdir/zshrc ~/.zshrc
@@ -37,8 +42,9 @@ ln -s $scriptdir/gitignore ~/.gitignore
 ln -s $scriptdir/tmux ~/.tmux.conf
 ln -s $scriptdir/gemrc ~/.gemrc
 ln -s $scriptdir/bash_profile ~/.bash_profile
-mkdir ~/.config/{i3,i3status}
-ln -s $scriptdir/i3config ~/.config/i3/config
+mkdir ~/.i3/
+mkdir ~/.config/i3status/
+ln -s $scriptdir/i3config ~/.i3/config
 ln -s $scriptdir/i3status ~/.config/i3status/config
 
 echo "Updating submodules"
